adding swagger to quarkus

+++ first-microservice-product/src/main/java/com/hector/producto/app/repositories/ProductRepository.java
@@ -6,6 +6,7 @@ import javax.enterprise.context.ApplicationScoped;
 import javax.persistence.EntityManager;
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Objects;
 
 @ApplicationScoped
 public class ProductRepository {
@@ -23,9 +24,14 @@ public class ProductRepository {
 
     @Transactional
     public void delete(Long id){
-        Producto producto = entityManager.createQuery("SELECT p FROM Producto p WHERE p.id = :id")
-                .setParameter("id", id)
-                .unwrap(Producto.class);
+//        Producto producto = entityManager.createQuery("SELECT p FROM Producto p WHERE p.id = :id")
+//                .setParameter("id", id)
+//                .unwrap(Producto.class);
+        Producto producto = entityManager.find(Producto.class, id);
+
+        if (Objects.isNull(producto))
+            throw new IllegalArgumentException("El producto con el id: "+id + " no existe");
+
         entityManager.remove(producto);
     }
